+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.75
 ### is.codion.swing.common.ui
 - ListComboBox refactored and simplified.
+- TemporalFieldPanel.supports() removed.
+- CalendarPanel.supportedTypes() replaced with supports(class).
 
 ## 0.18.74
 ### is.codion.common.utilities